fix: support multiple ISRC responses for tidal, improve logger
TIDAL also returns multiple songs for a given ISRC, so we should return
the first result. We didn't really document that we did that before, but
our custom TIDAL client does now.

I updated loggers to be aware of the provider that they are associated
with and passed them to the providers. Also updated the TIDAL API client
to log request information at the debug log level.